Refactor transaction status responses on a trait
